requirements.txt
openai
gitpython